Move FlipAxis to Geometry namespace
