+++ Assets/Tests/Geometry/Shapes/IsometricRectangle_Tests.cs
@@ -1,6 +1,7 @@
 using NUnit.Framework;
 
 using PAC.DataStructures;
+using PAC.Geometry;
 using PAC.Geometry.Shapes;
 using PAC.Tests.Geometry.Shapes.DefaultTests;
 using PAC.Tests.Geometry.Shapes.RequiredTests;
